Code share `dra` RunMainEngine
