+++ config/symbols.pspeu.dra.txt
@@ -769,7 +769,6 @@ D_801375C8 = 0x091CDCD8;
 g_canRevealMap = 0x091CDCE0;
 D_801375BC = 0x091CDCE8;
 D_801375C0 = 0x091CDCEC;
-D_801375C4 = 0x091CDCF0;
 g_ServantPrevious = 0x091CDD58;
 D_80137958 = 0x091CDD60;
 g_PrevEquippedWeapons = 0x091CDD68;
