+++ include/game.h
@@ -2054,8 +2054,7 @@ extern s32 g_ScrollDeltaY;
 extern s32 D_80097910;
 extern DemoMode g_DemoMode;
 extern s32 g_LoadOvlIdx; // 0x80097918
-extern s32 D_8009791C;
-extern s32 D_80097920;
+extern Point32 D_8009791C;
 extern s32 D_80097924;
 extern s32 D_80097928;
 extern GpuUsage g_GpuUsage;
