+++ src/pc/stubs.c
@@ -54,8 +54,7 @@ s32 g_ScrollDeltaX;
 s32 g_ScrollDeltaY;
 s32 D_80097910;
 s32 g_LoadOvlIdx;
-s32 D_8009791C;
-s32 D_80097920;
+Point32 D_8009791C;
 s32 D_80097924;
 s32 D_80097928;
 s32 D_8006C3AC;
